pyproject.toml
[project]
name = "transcript-api"
version = "0.1.0"
description = "Video to transcript service"
authors = [{ name = "Secil Sen" }]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.116.1,<0.117.0",
    "uvicorn>=0.35.0,<0.36.0",
    "google-generativeai>=0.8.0,<1.0.0",
    "pandas>=2.3.1,<3.0.0",
    "python-dotenv>=1.1.1,<2.0.0",
    "numpy>=1.21.0,<2.0",
    "scipy>=1.7.0",
    "librosa>=0.9.0",
    "pydub>=0.25.1,<0.26.0",
    "json5>=0.12.0,<0.13.0",
    "soundfile>=0.13.1,<0.14.0",
    "openai-whisper",
    "pyannote.audio>=3.1.1",
    "torch>=2.0.0",
    "torchaudio>=2.0.0",
    "aiohttp<3.9.0",
    "httpx>=0.24.0,<1.0.0",
    "rapidfuzz>=3.0.0,<4.0.0",
    "boto3>=1.28.0,<2.0.0"
]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "transcript-api"
version = "0.1.0"
description = "Video to transcript service"
authors = ["Secil Sen"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = ">=3.9"
fastapi = ">=0.116.1,<0.117.0"
uvicorn = ">=0.35.0,<0.36.0"
google-generativeai = ">=0.8.0,<1.0.0"
pandas = ">=2.3.1,<3.0.0"
python-dotenv = ">=1.1.1,<2.0.0"
numpy = ">=1.21.0,<2.0"
scipy = ">=1.7.0"
librosa = ">=0.9.0"
pydub = ">=0.25.1,<0.26.0"
json5 = ">=0.12.0,<0.13.0"
soundfile = ">=0.13.1,<0.14.0"
openai-whisper = "*"
"pyannote.audio" = ">=3.1.1"
torch = ">=2.0.0"
torchaudio = ">=2.0.0"
aiohttp = "<3.9.0"
httpx = ">=0.24.0,<1.0.0"
rapidfuzz = ">=3.0.0,<4.0.0"
boto3 = ">=1.28.0,<2.0.0"
matplotlib = ">=3.5.0"
google-cloud-aiplatform = ">=1.34.0,<2.0.0"
google-cloud-monitoring = ">=2.15.0,<3.0.0"
google-cloud-logging = ">=3.8.0,<4.0.0"
pydantic = ">=2.0.0,<3.0.0"
asyncio-mqtt = {version = ">=0.16.0,<1.0.0", optional = true}
replicate = "^1.0.7"

[tool.poetry.group.monitoring.dependencies]
google-cloud-aiplatform = ">=1.34.0,<2.0.0"
google-cloud-monitoring = ">=2.15.0,<3.0.0"
google-cloud-logging = ">=3.8.0,<4.0.0"
pydantic = ">=2.0.0,<3.0.0"

[tool.poetry.extras]
monitoring = ["google-cloud-aiplatform", "google-cloud-monitoring", "google-cloud-logging", "pydantic"]
